+++ plugin/src/test/java/com/stratio/cassandra/lucene/schema/mapping/GeoPointMapperTest.java
@@ -133,6 +133,15 @@ public class GeoPointMapperTest extends AbstractMapperTest {
     }
 
     @Test
+    public void testGetLatitudeFromNullColumn() {
+        GeoPointMapper mapper = geoPointMapper("lat", "lon").build("field");
+        Columns columns = new Columns();
+        columns.add(Column.builder("lat").composedValue(null, Int32Type.instance));
+        columns.add(Column.builder("lon").composedValue(0, Int32Type.instance));
+        assertNull("Latitude is not properly parsed", mapper.readLatitude(columns));
+    }
+
+    @Test
     public void testGetLatitudeFromIntColumn() {
         GeoPointMapper mapper = geoPointMapper("lat", "lon").build("field");
         Columns columns = new Columns();
@@ -235,6 +244,15 @@ public class GeoPointMapperTest extends AbstractMapperTest {
         columns.add(Column.builder("lat").composedValue(new Short("91"), ShortType.instance));
         columns.add(Column.builder("lon").composedValue(0, Int32Type.instance));
         mapper.readLatitude(columns);
+    }
+
+    @Test
+    public void testGetLongitudeFromNullColumn() {
+        GeoPointMapper mapper = geoPointMapper("lat", "lon").build("field");
+        Columns columns = new Columns();
+        columns.add(Column.builder("lat").composedValue(5, Int32Type.instance));
+        columns.add(Column.builder("lon").composedValue(null, Int32Type.instance));
+        assertNull("Longitude is not properly parsed", mapper.readLongitude(columns));
     }
 
     @Test